feat(console): split ops and tenant panels with org context bootstrap
