+++ backend/.env.example
@@ -145,7 +145,10 @@ FAP_ADMIN_TOKEN=
 EVENT_INGEST_TOKEN=
 FAP_ADMIN_PANEL_ENABLED=false
 FAP_ADMIN_GUARD=admin
-FAP_ADMIN_URL=http://localhost:18010/admin
+FAP_ADMIN_URL=http://localhost:18010/ops
+FAP_TENANT_PANEL_ENABLED=false
+FAP_TENANT_GUARD=tenant
+FAP_TENANT_URL=http://localhost:18010/tenant
 
 # =========================
 # Webhooks
