+++ backend/app/Filament/Ops/Resources/AdminUserResource/Pages/ListAdminUsers.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace App\Filament\Ops\Resources\AdminUserResource\Pages;
+
+use App\Filament\Ops\Resources\AdminUserResource;
+use Filament\Actions;
+use Filament\Resources\Pages\ListRecords;
+
+class ListAdminUsers extends ListRecords
+{
+    protected static string $resource = AdminUserResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [
+            Actions\CreateAction::make(),
+        ];
+    }
+}
