+++ backend/app/Filament/Ops/Resources/AuditLogResource/Pages/ListAuditLogs.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Filament\Ops\Resources\AuditLogResource\Pages;
+
+use App\Filament\Ops\Resources\AuditLogResource;
+use Filament\Resources\Pages\ListRecords;
+
+class ListAuditLogs extends ListRecords
+{
+    protected static string $resource = AuditLogResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [];
+    }
+}
