+++ backend/app/Filament/Ops/Resources/ContentReleaseResource/Pages/ListContentReleases.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Filament\Ops\Resources\ContentReleaseResource\Pages;
+
+use App\Filament\Ops\Resources\ContentReleaseResource;
+use Filament\Resources\Pages\ListRecords;
+
+class ListContentReleases extends ListRecords
+{
+    protected static string $resource = ContentReleaseResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [];
+    }
+}
