+++ backend/app/Filament/Ops/Resources/PermissionResource/Pages/CreatePermission.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace App\Filament\Ops\Resources\PermissionResource\Pages;
+
+use App\Filament\Ops\Resources\PermissionResource;
+use Filament\Resources\Pages\CreateRecord;
+
+class CreatePermission extends CreateRecord
+{
+    protected static string $resource = PermissionResource::class;
+
+    protected function afterCreate(): void
+    {
+        $record = $this->record;
+        app(\App\Services\Audit\AuditLogger::class)->log(
+            request(),
+            'permission_create',
+            'Permission',
+            (string) $record->id,
+            ['name' => $record->name]
+        );
+    }
+}
