+++ backend/app/Filament/Ops/Resources/PermissionResource/Pages/EditPermission.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace App\Filament\Ops\Resources\PermissionResource\Pages;
+
+use App\Filament\Ops\Resources\PermissionResource;
+use Filament\Actions;
+use Filament\Resources\Pages\EditRecord;
+
+class EditPermission extends EditRecord
+{
+    protected static string $resource = PermissionResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [
+            Actions\DeleteAction::make(),
+        ];
+    }
+
+    protected function afterSave(): void
+    {
+        $record = $this->record;
+        app(\App\Services\Audit\AuditLogger::class)->log(
+            request(),
+            'permission_update',
+            'Permission',
+            (string) $record->id,
+            ['name' => $record->name]
+        );
+    }
+}
