+++ backend/app/Filament/Ops/Resources/PermissionResource/Pages/ListPermissions.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace App\Filament\Ops\Resources\PermissionResource\Pages;
+
+use App\Filament\Ops\Resources\PermissionResource;
+use Filament\Actions;
+use Filament\Resources\Pages\ListRecords;
+
+class ListPermissions extends ListRecords
+{
+    protected static string $resource = PermissionResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [
+            Actions\CreateAction::make(),
+        ];
+    }
+}
