+++ backend/app/Filament/Ops/Resources/RoleResource/Pages/EditRole.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace App\Filament\Ops\Resources\RoleResource\Pages;
+
+use App\Filament\Ops\Resources\RoleResource;
+use Filament\Actions;
+use Filament\Resources\Pages\EditRecord;
+
+class EditRole extends EditRecord
+{
+    protected static string $resource = RoleResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [
+            Actions\DeleteAction::make(),
+        ];
+    }
+
+    protected function afterSave(): void
+    {
+        $record = $this->record;
+        app(\App\Services\Audit\AuditLogger::class)->log(
+            request(),
+            'role_update',
+            'Role',
+            (string) $record->id,
+            ['name' => $record->name]
+        );
+    }
+}
