+++ backend/app/Filament/Ops/Resources/RoleResource/Pages/ListRoles.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace App\Filament\Ops\Resources\RoleResource\Pages;
+
+use App\Filament\Ops\Resources\RoleResource;
+use Filament\Actions;
+use Filament\Resources\Pages\ListRecords;
+
+class ListRoles extends ListRecords
+{
+    protected static string $resource = RoleResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [
+            Actions\CreateAction::make(),
+        ];
+    }
+}
