--- backend/app/Filament/Resources/AdminUserResource/Pages/CreateAdminUser.php
@@ -1,23 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\AdminUserResource\Pages;
-
-use App\Filament\Resources\AdminUserResource;
-use Filament\Resources\Pages\CreateRecord;
-
-class CreateAdminUser extends CreateRecord
-{
-    protected static string $resource = AdminUserResource::class;
-
-    protected function afterCreate(): void
-    {
-        $record = $this->record;
-        app(\App\Services\Audit\AuditLogger::class)->log(
-            request(),
-            'admin_user_create',
-            'AdminUser',
-            (string) $record->id,
-            ['email' => $record->email]
-        );
-    }
-}
