--- backend/app/Filament/Resources/AuditLogResource/Pages/ListAuditLogs.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\AuditLogResource\Pages;
-
-use App\Filament\Resources\AuditLogResource;
-use Filament\Resources\Pages\ListRecords;
-
-class ListAuditLogs extends ListRecords
-{
-    protected static string $resource = AuditLogResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [];
-    }
-}
