--- backend/app/Filament/Resources/ContentReleaseResource/Pages/ListContentReleases.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\ContentReleaseResource\Pages;
-
-use App\Filament\Resources\ContentReleaseResource;
-use Filament\Resources\Pages\ListRecords;
-
-class ListContentReleases extends ListRecords
-{
-    protected static string $resource = ContentReleaseResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [];
-    }
-}
