--- backend/app/Filament/Resources/PermissionResource.php
@@ -1,84 +1,0 @@
-<?php
-
-namespace App\Filament\Resources;
-
-use App\Filament\Resources\PermissionResource\Pages;
-use App\Models\Permission;
-use App\Support\Rbac\PermissionNames;
-use Filament\Forms;
-use Filament\Forms\Form;
-use Filament\Resources\Resource;
-use Filament\Tables;
-use Filament\Tables\Table;
-
-class PermissionResource extends Resource
-{
-    protected static ?string $model = Permission::class;
-
-    protected static ?string $navigationIcon = 'heroicon-o-key';
-    protected static ?string $navigationGroup = 'Admin';
-    protected static ?string $navigationLabel = 'Permissions';
-
-    public static function canViewAny(): bool
-    {
-        $user = auth((string) config('admin.guard', 'admin'))->user();
-        return $user !== null
-            && method_exists($user, 'hasPermission')
-            && $user->hasPermission(PermissionNames::ADMIN_OWNER);
-    }
-
-    public static function canCreate(): bool
-    {
-        return self::canViewAny();
-    }
-
-    public static function canEdit($record): bool
-    {
-        return self::canViewAny();
-    }
-
-    public static function canDelete($record): bool
-    {
-        return self::canViewAny();
-    }
-
-    public static function form(Form $form): Form
-    {
-        return $form
-            ->schema([
-                Forms\Components\TextInput::make('name')
-                    ->required()
-                    ->maxLength(128),
-                Forms\Components\Textarea::make('description')
-                    ->rows(2)
-                    ->maxLength(255),
-            ]);
-    }
-
-    public static function table(Table $table): Table
-    {
-        return $table
-            ->columns([
-                Tables\Columns\TextColumn::make('name')->searchable(),
-                Tables\Columns\TextColumn::make('description')->limit(50),
-                Tables\Columns\TextColumn::make('created_at')->dateTime()->sortable(),
-            ])
-            ->actions([
-                Tables\Actions\EditAction::make(),
-            ])
-            ->bulkActions([
-                Tables\Actions\BulkActionGroup::make([
-                    Tables\Actions\DeleteBulkAction::make(),
-                ]),
-            ]);
-    }
-
-    public static function getPages(): array
-    {
-        return [
-            'index' => Pages\ListPermissions::route('/'),
-            'create' => Pages\CreatePermission::route('/create'),
-            'edit' => Pages\EditPermission::route('/{record}/edit'),
-        ];
-    }
-}
