--- backend/app/Filament/Resources/PermissionResource/Pages/ListPermissions.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\PermissionResource\Pages;
-
-use App\Filament\Resources\PermissionResource;
-use Filament\Actions;
-use Filament\Resources\Pages\ListRecords;
-
-class ListPermissions extends ListRecords
-{
-    protected static string $resource = PermissionResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            Actions\CreateAction::make(),
-        ];
-    }
-}
