--- backend/app/Filament/Resources/RoleResource.php
@@ -1,66 +1,0 @@
-<?php
-
-namespace App\Filament\Resources;
-
-use App\Filament\Resources\RoleResource\Pages;
-use App\Models\Role;
-use Filament\Forms;
-use Filament\Forms\Form;
-use Filament\Resources\Resource;
-use Filament\Tables;
-use Filament\Tables\Table;
-
-class RoleResource extends Resource
-{
-    protected static ?string $model = Role::class;
-
-    protected static ?string $navigationIcon = 'heroicon-o-shield-check';
-    protected static ?string $navigationGroup = 'Admin';
-    protected static ?string $navigationLabel = 'Roles';
-
-    public static function form(Form $form): Form
-    {
-        return $form
-            ->schema([
-                Forms\Components\TextInput::make('name')
-                    ->required()
-                    ->maxLength(64),
-                Forms\Components\Textarea::make('description')
-                    ->rows(3)
-                    ->maxLength(255),
-                Forms\Components\CheckboxList::make('permissions')
-                    ->relationship('permissions', 'name')
-                    ->columns(2),
-            ]);
-    }
-
-    public static function table(Table $table): Table
-    {
-        return $table
-            ->columns([
-                Tables\Columns\TextColumn::make('name')->searchable(),
-                Tables\Columns\TextColumn::make('description')->limit(50),
-                Tables\Columns\TextColumn::make('permissions_count')
-                    ->counts('permissions')
-                    ->label('Permissions'),
-                Tables\Columns\TextColumn::make('created_at')->dateTime()->sortable(),
-            ])
-            ->actions([
-                Tables\Actions\EditAction::make(),
-            ])
-            ->bulkActions([
-                Tables\Actions\BulkActionGroup::make([
-                    Tables\Actions\DeleteBulkAction::make(),
-                ]),
-            ]);
-    }
-
-    public static function getPages(): array
-    {
-        return [
-            'index' => Pages\ListRoles::route('/'),
-            'create' => Pages\CreateRole::route('/create'),
-            'edit' => Pages\EditRole::route('/{record}/edit'),
-        ];
-    }
-}
