--- backend/app/Filament/Resources/RoleResource/Pages/CreateRole.php
@@ -1,23 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\RoleResource\Pages;
-
-use App\Filament\Resources\RoleResource;
-use Filament\Resources\Pages\CreateRecord;
-
-class CreateRole extends CreateRecord
-{
-    protected static string $resource = RoleResource::class;
-
-    protected function afterCreate(): void
-    {
-        $record = $this->record;
-        app(\App\Services\Audit\AuditLogger::class)->log(
-            request(),
-            'role_create',
-            'Role',
-            (string) $record->id,
-            ['name' => $record->name]
-        );
-    }
-}
