--- backend/app/Filament/Resources/RoleResource/Pages/ListRoles.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\RoleResource\Pages;
-
-use App\Filament\Resources\RoleResource;
-use Filament\Actions;
-use Filament\Resources\Pages\ListRecords;
-
-class ListRoles extends ListRecords
-{
-    protected static string $resource = RoleResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            Actions\CreateAction::make(),
-        ];
-    }
-}
