+++ backend/app/Filament/Tenant/Pages/TenantDashboard.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Filament\Tenant\Pages;
+
+use Filament\Pages\Dashboard;
+
+class TenantDashboard extends Dashboard
+{
+    protected static ?string $title = 'Tenant Console';
+}
