+++ backend/bootstrap/providers.php
@@ -4,7 +4,10 @@ $filamentAvailable = class_exists(\Filament\PanelProvider::class);
 
 return array_values(array_filter([
     App\Providers\AppServiceProvider::class,
-    (env('FAP_ADMIN_PANEL_ENABLED', true) && $filamentAvailable)
+    (env('FAP_ADMIN_PANEL_ENABLED', false) && $filamentAvailable)
         ? App\Providers\Filament\AdminPanelProvider::class
+        : null,
+    (env('FAP_TENANT_PANEL_ENABLED', false) && $filamentAvailable)
+        ? App\Providers\Filament\TenantPanelProvider::class
         : null,
 ]));
