+++ backend/config/admin.php
@@ -4,9 +4,9 @@ use App\Support\Rbac\PermissionNames;
 
 return [
     'token' => env('FAP_ADMIN_TOKEN', ''),
-    'panel_enabled' => (bool) env('FAP_ADMIN_PANEL_ENABLED', true),
+    'panel_enabled' => (bool) env('FAP_ADMIN_PANEL_ENABLED', false),
     'guard' => env('FAP_ADMIN_GUARD', 'admin'),
-    'url' => env('FAP_ADMIN_URL', 'http://localhost:18010/admin'),
+    'url' => env('FAP_ADMIN_URL', 'http://localhost:18010/ops'),
 
     'permissions' => PermissionNames::all(),
     'role_permissions' => PermissionNames::defaultRolePermissions(),
